+++ config/cst.py
@@ -7,6 +7,8 @@ START_EVAL_PERTINENCE = 1
 SOCIAL_SPECIFIC_CONFIG_PATH = "config/social_evaluator_config/"
 SOCIAL_CONFIG_REFRESH_RATE = "refresh_rate_seconds"
 
+SOCIAL_EVALUATOR_NOT_THREADED_UPDATE_RATE = 1
+
 
 class EvaluatorRisk(Enum):
     LOW = 1
